# telehealth-bangladesh/API's/telehealth_openapi_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HealNSight Telehealth Bangladesh API",
    "description": "Complete OpenAPI 3.0.3 specification covering all 58 endpoints across Django REST Framework, Node.js Auth Microservice, and FastAPI Worker.",
    "version": "1.0.0",
    "contact": {
      "name": "HealNSight Technical Team",
      "url": "https://github.com/infectedsourav/telehealth-bangladesh"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Django REST Backend (Core Clinical & Operations)"
    },
    {
      "url": "http://localhost:5000",
      "description": "Node.js Auth & OTP Microservice"
    },
    {
      "url": "http://localhost:6000",
      "description": "FastAPI AI & Vision Worker"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token (obtainable from /api/login/ or /api/auth/login)."
      }
    }
  },
  "paths": {
...
    "/api/login/": {
      "post": {
        "tags": [
          "Django Auth"
        ],
        "summary": "Obtain JWT Token Pair",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "username": "kazi_patient",
                  "password": "Password123!"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payment/initiate/": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Initiate SSLCommerz Gateway Session",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "amount": 1000.0,
                  "consultation_id": 1,
                  "patient_name": "Kazi",
                  "patient_phone": "+8801700000001"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payment/ipn/": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "SSLCommerz Instant Payment Notification",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "tran_id": "TXN_123",
                  "status": "VALIDATED"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment/status/{transaction_ref}/": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Query Payment Status",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  }
}